[server] Moved QgsFcgiServerRequest into its own header and cpp file
