+++ src/server/CMakeLists.txt
@@ -48,6 +48,7 @@ SET(qgis_mapserv_SRCS
   qgsserverrequest.cpp
   qgsserverresponse.cpp
   qgsfcgiserverresponse.cpp
+  qgsfcgiserverrequest.cpp
   qgsbufferserverresponse.cpp
   qgsbufferserverrequest.cpp
   qgsfilterresponsedecorator.cpp
